+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+<a name="1.1.0"></a> ### 1.1.0 (2022-12-12)
+
+#### Features
+- Added check against duplication of old passwords
+- Added config params for password min/max length and old passwords validation
+
+
 <a name="1.0.0"></a> ### 1.0.0 (2022-04-29)
 
 #### Features